refactor(linux): make room under the Chuyển mã panes
Batch mode had grown to 143 of the 150 lines `scripts/check-loc.sh` allows,
which is not enough room to hang another control off the header. Lift
`files::Pane::refresh` into `files/show.rs` so batch mode has the same shape
text mode already has: the pane file holds the skeleton and the wiring, and
the redraw lives next door.

Promote `click` out of `text/footer.rs` into `widget.rs` at the same time,
widened from a `fn` pointer to `impl Fn` so a caller can capture. The footer
only ever pointed it at a free function, but the casing chips landing next
each need to remember their own position in the menu, and a second copy of
the downgrade/upgrade dance is exactly what `widget.rs` exists to prevent.

No behaviour change.
